+++ SpaceGamePrelim/MapCell.cpp
@@ -11,7 +11,7 @@ MapCell::MapCell()
 MapCell::MapCell(std::vector<std::string> TextureIDs, TextureProperties* Properties, MapCoordinate Position)
 {
 
-	for (int i = 0; i < TextureIDs.size(); i++)
+	for (size_t i = 0; i < TextureIDs.size(); i++)
 	{
 		m_TextureIDs.push_back(TextureIDs.back());
 		TextureIDs.pop_back();
